Refactor epic ctas
