+++ dotcom-rendering/src/components/marketing/epics/ContributionsLiveblogEpic.tsx
@@ -19,7 +19,7 @@ import type {
 	EpicProps,
 	Tracking,
 } from '@guardian/support-dotcom-components/dist/shared/src/types';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useIsInView } from '../../../lib/useIsInView';
 import type { ReactComponent } from '../lib/ReactComponent';
 import { replaceArticleCount } from '../lib/replaceArticleCount';
@@ -28,10 +28,8 @@ import {
 	createViewEventFromTracking,
 } from '../lib/tracking';
 import { logEpicView } from '../lib/viewLog';
-import { ContributionsEpicCtas } from './ContributionsEpicCtas';
 import { ContributionsEpicNewsletterSignup } from './ContributionsEpicNewsletterSignup';
-import { ThreeTierChoiceCards } from './ThreeTierChoiceCards';
-import { getDefaultThreeTierAmount } from './utils/threeTierChoiceCardAmounts';
+import { ContributionsEpicCtasContainer } from './ctas/ContributionsEpicCtasContainer';
 
 // Hard-coded AB TEST - picking up ab test name and variant name from the tracking object
 // then applying a different colour if it matches, or the default colour if it doesn't.
@@ -156,12 +154,6 @@ export const ContributionsLiveblogEpic: ReactComponent<EpicProps> = ({
 }: EpicProps): JSX.Element => {
 	const { newsletterSignup } = variant;
 
-	const isNonVatCompliantCountry =
-		variant.choiceCardAmounts?.testName === 'VAT_COMPLIANCE';
-
-	const showChoiceCards =
-		variant.showChoiceCards && !isNonVatCompliantCountry;
-
 	const [hasBeenSeen, setNode] = useIsInView({
 		debounce: true,
 	});
@@ -198,23 +190,12 @@ export const ContributionsLiveblogEpic: ReactComponent<EpicProps> = ({
 		replaceNonArticleCountPlaceholders(variant.heading) ||
 		'Support the Guardian';
 
-	const defaultThreeTierAmount = getDefaultThreeTierAmount(countryCode);
-	const [
-		threeTierChoiceCardSelectedAmount,
-		setThreeTierChoiceCardSelectedAmount,
-	] = useState<number>(defaultThreeTierAmount);
-
 	if (
 		cleanParagraphs.some(containsNonArticleCountPlaceholder) ||
 		containsNonArticleCountPlaceholder(cleanHeading)
 	) {
 		return <></>;
 	}
-
-	const variantOfChoiceCard =
-		countryCode === 'US'
-			? 'US_THREE_TIER_CHOICE_CARDS'
-			: 'THREE_TIER_CHOICE_CARDS';
 
 	return (
 		<div data-testid="contributions-liveblog-epic" ref={setNode}>
@@ -235,34 +216,15 @@ export const ContributionsLiveblogEpic: ReactComponent<EpicProps> = ({
 						tracking={tracking}
 					/>
 				) : (
-					<>
-						{showChoiceCards && (
-							<ThreeTierChoiceCards
-								countryCode={countryCode}
-								selectedAmount={
-									threeTierChoiceCardSelectedAmount
-								}
-								setSelectedAmount={
-									setThreeTierChoiceCardSelectedAmount
-								}
-								variantOfChoiceCard={variantOfChoiceCard}
-							/>
-						)}
-						<ContributionsEpicCtas
-							variant={variant}
-							tracking={tracking}
-							countryCode={countryCode}
-							articleCounts={articleCounts}
-							onReminderOpen={onReminderOpen}
-							fetchEmail={fetchEmail}
-							submitComponentEvent={submitComponentEvent}
-							showChoiceCards={showChoiceCards}
-							threeTierChoiceCardSelectedAmount={
-								threeTierChoiceCardSelectedAmount
-							}
-							variantOfChoiceCard={variantOfChoiceCard}
-						/>
-					</>
+					<ContributionsEpicCtasContainer
+						variant={variant}
+						tracking={tracking}
+						countryCode={countryCode}
+						articleCounts={articleCounts}
+						onReminderOpen={onReminderOpen}
+						fetchEmail={fetchEmail}
+						submitComponentEvent={submitComponentEvent}
+					/>
 				)}
 			</section>
 		</div>
