+++ dotcom-rendering/src/components/marketing/epics/ctas/ContributionsEpicChoiceCards.tsx
@@ -14,10 +14,10 @@ import type {
 	SelectedAmountsVariant,
 } from '@guardian/support-dotcom-components/dist/shared/src/types';
 import { useEffect } from 'react';
-import { useIsInView } from '../../../lib/useIsInView';
-import { contributionType } from '../lib/choiceCards';
-import type { ChoiceCardSelection } from '../lib/choiceCards';
-import type { ReactComponent } from '../lib/ReactComponent';
+import { useIsInView } from '../../../../lib/useIsInView';
+import { contributionType } from '../../lib/choiceCards';
+import type { ChoiceCardSelection } from '../../lib/choiceCards';
+import type { ReactComponent } from '../../lib/ReactComponent';
 
 // CSS Styling
 // -------------------------------------------
