+++ dotcom-rendering/src/components/marketing/epics/ctas/ContributionsEpicReminderSignedIn.tsx
@@ -19,9 +19,9 @@ import {
 	SvgCross,
 } from '@guardian/source/react-components';
 import { StraightLines } from '@guardian/source-development-kitchen/react-components';
-import { Hide } from '../../Hide';
-import type { ReactComponent } from '../lib/ReactComponent';
-import { ensureHasPreposition, ReminderStatus } from '../lib/reminders';
+import { Hide } from '../../../Hide';
+import type { ReactComponent } from '../../lib/ReactComponent';
+import { ensureHasPreposition, ReminderStatus } from '../../lib/reminders';
 
 // --- Styles --- //
